Create an ExternalGame component
